requirements.txt
# Core data stack
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0
openpyxl==3.1.2
python-dateutil==2.9.0.post0

# Spark & Parquet 
pyspark==3.5.1       

# Machine learning
xgboost==2.0.3        
scikit-learn==1.5.0   
scipy==1.13.1        

# Helpers
setuptools==69.2.0
wheel==0.42.0
